Mediabuilder refactoring
Added BUNDLE_MEDIA setting.  This will control if we are want media to
be combined, compressed, and moved to static storage.

Added nicer interface for the settings dict

Moved source_urls.py to urls.py.  Made it so you can include it at all
times, but it will only have urls in it if BUNDLE_MEDIA is True

Moved .gitignore stuff to the mediabuilder app