Added small screen nav menu
The menu gets toggled on/off by clicking a button.  Made a spiffy
transition for the menu icon -- hopefully that scss file can be
re-used.

Some reworking of the HTML/CSS for the nav menu.  Now everything is in
1 ul.

Added jquery to the downloads.

Made mediabuilder throw an exception if a JS glob doesn't match
anything.